# api/swagger.yml
openapi: 3.0.1
info:
  title: lakeFS API
  description: lakeFS HTTP API
  version: 0.1.0
servers:
- url: /api/v1
security:
- jwt_token: []
- basic_auth: []
paths:
...
  /repositories/{repository}/refs/{ref}/objects:
    get:
      tags:
      - objects
      summary: get object content
      operationId: getObject
      parameters:
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: ref
        in: path
        description: a reference (could be either a branch or a commit ID)
        required: true
        schema:
          type: string
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        200:
          description: object content
          headers:
            ETag:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
            Content-Length:
              schema:
                type: integer
                format: int64
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: resource not found
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/error'
        410:
          description: object expired
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: generic error response
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/error'
  /repositories/{repository}/branches/{branch}/objects:
    put:
      tags:
      - objects
      summary: stage an object's metadata for the given branch
      operationId: stageObject
      parameters:
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: branch
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/object_stage_creation'
        required: false
      responses:
        201:
          description: object metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/object_stats'
        400:
          description: invalid object metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: repository or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-codegen-request-body-name: object
    post:
      tags:
      - objects
      summary: upload object content
      operationId: uploadObject
      parameters:
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: branch
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: query
        required: true
        schema:
          type: string
      - name: storageClass
        in: query
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                content:
                  type: string
                  description: Object content to upload
                  format: binary
      responses:
        201:
          description: object metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/object_stats'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: repository or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
      - objects
      summary: delete object
      operationId: deleteObject
      parameters:
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: branch
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        204:
          description: object deleted successfully
          content: {}
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: path or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /repositories/{repository}/refs/{ref}/objects/ls:
    get:
      tags:
      - objects
      summary: list objects under a given prefix
      operationId: listObjects
      parameters:
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: ref
        in: path
        description: a reference (could be either a branch or a commit ID)
        required: true
        schema:
          type: string
      - name: prefix
        in: query
        schema:
          type: string
      - name: after
        in: query
        schema:
          type: string
      - name: amount
        in: query
        schema:
          type: integer
          default: 100
      - name: delimiter
        in: query
        allowEmptyValue: true
        schema:
          type: string
          default: /
      responses:
        200:
          description: entry list
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: '#/components/schemas/pagination'
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/object_stats'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: prefix or branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
    pagination:
      required:
      - has_more
      type: object
      properties:
        has_more:
          type: boolean
        next_offset:
          type: string
        results:
          minimum: 0
          type: integer
        max_per_page:
          minimum: 0
          type: integer
...
    object_stats:
      type: object
      properties:
        path:
          type: string
        physical_address:
          type: string
        checksum:
          type: string
        mtime:
          type: integer
          format: int64
        size_bytes:
          type: integer
          format: int64
          nullable: true
        path_type:
          type: string
          enum:
          - common_prefix
          - object
    object_stage_creation:
      required:
      - checksum
      - physical_address
      - size_bytes
      type: object
      properties:
        physical_address:
          type: string
        checksum:
          type: string
        size_bytes:
          type: integer
          format: int64
        metadata:
          type: object
          additionalProperties:
            type: string
...
    error:
      type: object
      properties:
        message:
          type: string
          description: short message explaining the error
...
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
    jwt_token:
      type: apiKey
      name: X-JWT-Authorization
      in: header